+++ kubejs/server_scripts/recipes/enigmatica/replace_input.js
@@ -18,9 +18,18 @@ ServerEvents.recipes((event) => {
             replace_with: '#c:dyes/black'
         },
         {
-            filter: [{ mod: 'ae2', not: { output: 'ae2:cable_anchor' } }, { mod: 'ae2netanalyser' }],
+            filter: [
+                { mod: 'ae2', not: { output: 'ae2:cable_anchor' } },
+                { mod: 'ae2netanalyser' },
+                { output: /tackle_box/ }
+            ],
             to_replace: '#c:ingots/iron',
             replace_with: Ingredient.of(['#c:ingots/iron', '#c:ingots/tin'])
+        },
+        {
+            filter: [{ output: /tackle_box/ }],
+            to_replace: '#c:storage_blocks/iron',
+            replace_with: Ingredient.of(['#c:storage_blocks/iron', '#c:storage_blocks/tin'])
         },
         {
             filter: { mod: 'functionalstorage' },
@@ -28,7 +37,7 @@ ServerEvents.recipes((event) => {
             replace_with: 'framedblocks:framed_chest'
         },
         {
-            filter: { mod: 'handcrafted' },
+            filter: [{ mod: 'handcrafted' }, { mod: 'aquaculture' }],
             to_replace: '#c:chests',
             replace_with: '#c:chests/wooden'
         }